Update project setup instructions
Simplified the README setup instructions by replacing the local Python, Django, Redis, and MySQL installations with Docker. This change was made in order to improve the onboarding experience for new developers by minimizing the amount of configuration required to get the project up and running. The logs are also included for debugging purposes.